% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjogren_functions.R
\name{add_BCR}
\alias{add_BCR}
\title{Add VDJ data to seurat objects in a list}
\usage{
add_BCR(list, contigs)
}
\arguments{
\item{list}{A named list of seurat objects. This can be generated by the generate_seurat_objects() function.}

\item{contigs}{A named list of contigs from VDJ data. This list should be in the same order as the list of seurat objects.}
}
\value{
A named list of Seurat objects with VDJ data added.
}
\description{
This function adds VDJ data to a list of seurat objects using scRepertoire. The clonecall is set to "strict".
}
\examples{
list <- list(expression_matrix1, expression_matrix2)
          names(list) <- c("object1", "object2")
          seurat_list <- generate_seurat_objects(list)
         contig_list <- list(contig_1, contig_2)
         names(contig_list) <- c("object1", "object2")
          seurat_list <- add_BCR(seurat_list, contig_list)
}
\keyword{screpertoire,}
\keyword{seurat,}
\keyword{vdj}
