% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjogren_functions.R
\name{add_dotplot_annot}
\alias{add_dotplot_annot}
\title{Add dotplot annotation}
\usage{
add_dotplot_annot(seurat, mode = "cluster", include_IGHA_IGHD = F)
}
\arguments{
\item{seurat}{A seurat object with the following columns in its metadata: annotation: the cluster names or grouped_annotation: the grouped annotation generated by group_annot(),
group: the group the cells belong to (RF, ANA, both or NA), c_gene: the isotype as predicted by cellranger or a similar program}

\item{mode}{Which annotation to use: the cluster annotation ("cluster") or the grouped annotation ("group") generated by the group_annot() function. Default is cluster.}

\item{include_IGHA_IGHD}{logical indicating whether cells with an IGHA or IGHD isotype should be annotated too.}
}
\value{
A new column in the seurat metadata (dotplot_annot_res) giving the new annotation.
}
\description{
This function adds an annotation that can be used to separate groups in a dotplot. Cells with no autoreactivity assigned will be annotated based on their isotype,
and cells that are assigned an autoreactivity based on their group (RF, ANA or both). The amount of cells that are present within a group(cluster+ isotype/reactivity)
}
\examples{
seurat <- add_dotplot_annot(seurat, mode = "cluster", include_IGHA_IGHD = T)

}
\keyword{dotplot}
\keyword{seurat,}
