% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjogren_functions.R
\name{add_top_level}
\alias{add_top_level}
\title{Group Azimuth l2 annotations}
\usage{
add_top_level(seurat_object)
}
\arguments{
\item{seurat_object}{A seurat object on which RunAzimuth() has been called.}
}
\value{
A new column in the metadata of the seurat object, called top_level_annot, which contains the grouped annotations.
}
\description{
This function groups Azimuth level 2 annotations into larger, overarching groups. This is mainly meant for seurat objects consisting of of B cells.
This function should be run prior to normalisation.
}
\examples{
seurat <- RunAzimuth(seurat, ref = "tonsilref")
          seurat <- add_top_level(seurat)

}
\keyword{Azimuth}
\keyword{seurat,}
