% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjogren_functions.R
\name{annotate_clusters}
\alias{annotate_clusters}
\title{Annotate GO and KEGG pathways of all clusters}
\usage{
annotate_clusters(markers)
}
\arguments{
\item{list}{A named list of seurat objects. This can be generated by the generate_seurat_objects() function.}

\item{seurat_list}{A named list of seurat objects, which can be generated by the generate_seurat_objects() function.}
}
\value{
A list with a GO and a KEGG result for each cluster. The dotplot() function of clusterProfiler can be used on each list element.
}
\description{
This function performs GO and KEGG pathway analysis (using clusterProfiler) for each cluster based on the marker genes found with seurats FindAllMarkers() function. It does not discriminate between upregulated and downregulated markers.
If this distinction is desired, the data frame should be split and the function should be ran on each dataframe.
If gene names appear as their non-standard names, the function will try to convert these to their standard names.
}
\examples{
markers <- FindAllMarkers(seurat)
          cluster_annotation <- annotate_clusters(markers)
          dotplot(cluster_annotation[["GO"]])
}
\keyword{GO,}
\keyword{KEGG,}
\keyword{markers}
