% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjogren_functions.R
\name{assign_autoreactivity}
\alias{assign_autoreactivity}
\title{Assign autoreactivity}
\usage{
assign_autoreactivity(contig, sequences)
}
\arguments{
\item{contig}{The clonotyped and tested VDJ contigs as is the result from clonotype() and test.result().}

\item{sequences}{The amino acid sequences of B cells that are known to be autoreactive. Format is data.frame. This data.frame should also contain which donor the sequence was found in, and a name for the specific reactivity (for example, CARAAA-Ro52)}
}
\value{
The original contig file, with two new columns: their reactivities (named the same as in the sequences data.frame) and their antigenicity (for example, RF or La).
}
\description{
This function assigns cells as being autoreactive or not based on the CDR3 amino acid sequences of known autoreactive cells.
If the CDR3 aa sequence of a cell matches a known autoreactive sequence, the cell and its entire clonotype are annotated as autoreactive.
}
\examples{
contig <- clonotype(contig, threshold = 2, levenshtein = F)
            contig <- assign_autoreactivity(contig, sequences)
}
\keyword{autoreactivity}
