% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjogren_functions.R
\name{clonotype}
\alias{clonotype}
\title{Clonotyping function}
\usage{
clonotype(contig, threshold, levenshtein = F)
}
\arguments{
\item{contig}{The VDJ contigs (as a data frame, containing at least a barcode, VDJ assignments, and CDR3 nucleotide sequences).}

\item{threshold}{The allowed number of mismatches (as an integer) in the CDR3 region between two sequences in a clonotype.}

\item{levenshtein}{Should the distance be calculated based on the levenshtein distance (allowing indels)? Defaults to FALSE (using point mutation mismatches instead)}
}
\value{
The original contig file, with clonotype numbers for each cell in the "clonotype_new" column.
}
\description{
This function clonotypes VDJ contig files based on identical heavy chain VDJ assignments, identical light chain VJ assignments,
#and a user-defined threhold for the amount of nucleotide mismatches in the CDR3 region.
}
\examples{
contig <- clonotype(contig, threshold = 2, levenshtein = F)
}
\keyword{clonotyping}
