% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjogren_functions.R
\name{extract_tree_sequences}
\alias{extract_tree_sequences}
\title{Extract all barcode labels for each tree in a list}
\usage{
extract_tree_sequences(db)
}
\arguments{
\item{db}{A phylogenetic tree in format "phylo" that contain trees of multiple cellular barcodes.}
}
\value{
A list, named with the same numbers as the tree numbers, of cell barcode vectors.
}
\description{
This function makes a list of all cell barcodes that are present in a phylogenetic tree.
}
\examples{
db <- readIghyml(".....igphyml-pass.tab", format = "phylo")
db_list <- extract_tree_sequences(db)

}
\keyword{barcodes}
\keyword{phylogenetic}
\keyword{tree,}
