% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjogren_functions.R
\name{filtering_AIRR}
\alias{filtering_AIRR}
\title{Filter out cells with abberant IGH chains}
\usage{
filtering_AIRR(data)
}
\arguments{
\item{data}{An AIRR-formatted data frame containing light chain and heavy chain entries for each cell, describing the genes they use, their CDR3, etc.}
}
\value{
The same dataframe, but now with the mentioned cells filtered out. It also produces output on how many cells are removed.
}
\description{
This function filters out cells with two IGH chains or no IGH chain from an AIRR-formatted data frame.
}
\examples{
data <- filtering_AIRR(data)

}
\keyword{filtering,}
\keyword{immcantation}
\keyword{vdj,}
