% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjogren_functions.R
\name{find_autoreactive_clones}
\alias{find_autoreactive_clones}
\title{Find autoreactive cells within a list of phylogenetic trees}
\usage{
find_autoreactive_clones(
  db_seq,
  donor,
  data_merged,
  autoreactive_cells,
  seurat
)
}
\arguments{
\item{db_seq}{A list of cellular barcodes in every tree in a list, as can be generated by the extract_tree_sequences() function.}

\item{donor}{A character vector indicating the donor of which the autoreactive cell barcodes should be used. Options are D001, D002, D003, D004LN, D004PG, D007 and AIM.}

\item{data_merged}{A data.frame containing all cell barcodes (of every donor) and their corresponding contig data.}

\item{autoreactive_cells}{A data.frame stating the barcode of each autoreactive cell and their donor of origin}

\item{seurat}{A seurat object containing the cluster annotation of each cell in the metadata column "annotation".}
}
\value{
A list of autoreactive cell barcodes, their cluster annotation, and the tree number they were found in.
}
\description{
This function identifies all trees containing autoreactive cells, provided that a data.frame of autoreactive cell barcodes is given.
It also annotates every tree based on the metadata of the corresponding Seurat object.
}
\examples{
db_seq <- extract_tree_sequences(db)
clones_1 <- find_autoreactive_clones(db_seq, donor = "D001", data_merged, autoreactive_cells, seurat)

}
\keyword{autoreactivity}
\keyword{phylogenetic}
\keyword{tree,}
