% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjogren_functions.R
\name{generate_seurat_objects}
\alias{generate_seurat_objects}
\title{Generate seurat objects from a list}
\usage{
generate_seurat_objects(expression_matrices, min.cells = 3, min.genes = 200)
}
\arguments{
\item{expression_matrices}{A named list of expression matrices. These expression matrices can be, for example, cellranger outputs.}

\item{min.cells}{The minimum amount of cells a gene needs to be expressed in to be kept in the seurat object}

\item{min.genes}{The minimum amount of genes a cell needs to express to be kept in the seurat object}
}
\value{
A named list of Seurat objects with RNA and CITE assays (wherever applicable)
}
\description{
This is a simple function to generate a list of seurat objects from a list of count matrices containing gene expression data or gene expression data and CITE-seq data.
}
\examples{
list <- list(expression_matrix1, expression_matrix2)
          names(list) <- c("object1", "object2")
          seurat_list <- generate_seurat_objects(list)

}
\keyword{seurat}
