% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjogren_functions.R
\name{group_annot}
\alias{group_annot}
\title{Group clusters into larger groups}
\usage{
group_annot(seurat, annotation, mbc_index, gc_index, plasma_index, nbc_index)
}
\arguments{
\item{seurat}{A seurat object with the following columns in its metadata: annotation: the cluster names, tissue: the tissue of origin of that cell}

\item{annotation}{A vector with the unique annotation for each cluster. This can, for example, be the vector you use to rename the seurat clusters.}

\item{mbc_index}{A vector of the cluster number of all clusters belonging to the memory B cell (MBC) class.}

\item{gc_index}{A vector of the cluster number of all clusters belonging to the germinal center (GC) class.}

\item{plasma_index}{A vector of the cluster number of all clusters belonging to the plasma cell (PC) class.}

\item{nbc_index}{A vector of the cluster number of all clusters belonging to the naive B cell cell (NMC) class.}
}
\value{
A new column called grouped_annotation in the seurat metadata giving the new annotation.
}
\description{
Add an annotation that groups your Seurat clusters into 3 groups (MBC, PC, GC) per tissue (SG, LN, PBMC). Any cluster that does not belong to one of the three groups is assigned "Other".
The function throws an error if a cluster is assigned to multiple groups, so make sure no cluster number appears in multiple vectors. The cluster numbering starts counting from zero.
}
\examples{
annotation <- c("Germinal center", "CD27+ MBC", "CXCR4+ MBC", "IGHM+ PC", "IGHG+ PC")
seurat <- group_annot(seurat, annotation, gc_index = c(0), mbc_index = c(1,2), plasma_index = c(3,4))

}
\keyword{cluster}
\keyword{grouping,}
\keyword{seurat,}
