% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjogren_functions.R
\name{phylo_tree}
\alias{phylo_tree}
\title{Generate a phylogenetic tree plot}
\usage{
phylo_tree(tree_number, seurat = NULL, ident, annotation = T)
}
\arguments{
\item{tree_number}{A character of the tree number identifier (as given in the names of the tree list) that is to be plotted.}

\item{seurat}{A seurat object with the cluster annotation of each cell in a metadata column called "annotation".}

\item{ident}{The donor ident to be used when determining if the sequences are autoreactive or not.}
}
\value{
A phylogenetic tree plot of the indicated tree. The cluster identities for each cell are plotted as a bar to the side. A legend for this cluster annotation is added in the upper left corner.
A cluster identity of NA means that the cell was not found back in the suerat object.
The cells are coloured based on their autoreactivity: black if non-autoreactive, green in RF, pink if Ro60, gold if Ro52, purple if both Ro52 and Ro60, red if La, and cyan if the cell is autoreactive but does not match any of these combinations.
}
\description{
This function generates a plot of a phylogenetic tree, indicating any possible autoreactive sequences in the tree, and also adding the cluster annotation from Seurat.
For each donor, a corresponding db object and autoreactive (auto) data frame should exist. For example, for D001, a phylo tree named db1 should exist, as well as auto1, which is the result of running the find_autoreactive_clones() function for D001.
}
\examples{
plot <- phylo_tree("123_1", ident = "D003", seurat)

}
\keyword{autoreactivity,}
\keyword{phylogenetic}
\keyword{plot}
\keyword{tree,}
