% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjogren_functions.R
\name{plot_clone}
\alias{plot_clone}
\title{Make a UMAP of a specific clone}
\usage{
plot_clone(seurat, clone_number)
}
\arguments{
\item{seurat}{A seurat object which has already been UMAP reduced and clustered, and has a metadata column indicating the clonotype of each cell}

\item{clone_number}{A character (vector) indicating the clonotype that needs to be highlighted on the UMAP}
}
\value{
A UMAP with the indicated clonotypes highlighted.
}
\description{
This function produces a UMAP on which a specific clonotype is highlighted.
}
\examples{
plot <- plot_clone(seurat, c("1", "2"))

}
\keyword{clonotype}
\keyword{dimplot,}
\keyword{seurat,}
\keyword{umap,}
