% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjogren_functions.R
\name{remove_genes}
\alias{remove_genes}
\title{Remove contaminating genes}
\usage{
remove_genes(seurat_list)
}
\arguments{
\item{seurat_list}{A named list of seurat objects, which can be generated by the generate_seurat_objects() function.}

\item{list}{A named list of seurat objects. This can be generated by the generate_seurat_objects() function.}
}
\value{
A named list of Seurat objects with contaminating and non-interesting genes removed.
}
\description{
This function removes contaminating genes and genes that are not of interest (mitochondrial, ribosomal, MALAT1, GAPDH, microRNAs, ncRNAs) from the RNA assays of a list of seurat objects.
This function should be run prior to normalisation.
}
\examples{
list <- list(expression_matrix1, expression_matrix2)
          names(list) <- c("object1", "object2")
          seurat_list <- generate_seurat_objects(list)
          seurat_list <- remove_genes(seurat_list)
}
\keyword{RNA}
\keyword{seurat,}
