% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjogren_functions.R
\name{remove_vdj_doublets}
\alias{remove_vdj_doublets}
\title{Remove VDJ doublets}
\usage{
remove_vdj_doublets(matrix, vdj_b, vdj_t)
}
\arguments{
\item{matrix}{A matrix containing gene expression counts, or a list consisting of a gene expression count matrix and a CITE-seq count matrix. This function is directly applicable to 10X expression matrices generated by cellranger.}

\item{vdj_b}{A data.frame containing the VDJ B contig data}

\item{vdj_t}{A data.frame containing the VDJ T contig data}
}
\value{
The original count matrices with the doublet cells removed.
}
\description{
This function removes doublets from gene expression matrices (and CITE-seq matrices, if applicable) based on the VDJ data of that matrix.
Dual IGH or TRB chains are very rare in lymphocytes and are likely the result of a doublet. Also, cells should either express a BCR or a TCR, not both.
This function removes cells with dual IGH or TRB chains, or cells that appear in both the VDJ B and VDJ T data from the gene expression and CITE count matrices.
}
\examples{
matrix <- remove_vdj_doublets(matrix, vdj_b, vdj_t)
}
\keyword{doublet,}
\keyword{vdj}
