% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjogren_functions.R
\name{test.result}
\alias{test.result}
\title{Test whether the clonotype() function ran correctly}
\usage{
test.result(contig, threshold, levenshtein = F)
}
\arguments{
\item{contig}{The clonotyped VDJ contigs, which is the result of the clonotype() function..}

\item{threshold}{The allowed number of mismatches (as an integer) in the CDR3 region between two sequences in a clonotype. This should be the same threshold as used in the clonotype() function.}

\item{levenshtein}{Should the levenshtein distance be used in the checks? Set this to be equal to the clonotype() function. Default is FALSE.}
}
\value{
Prints out the amount of non-clonotyped cells, the number of cells with multiple light chain assignments, and whether the object passed the heavy chain and CDR3 checks.
}
\description{
This function tests whether the clonotypes generated by clonotype() really have identical heavy chain VDJ assignments and a maximum of nucleotide mismatches in the CDR3 given by the threshold.
It also shows how many cells were not assigned a clonotype, and how many cellss have multiple light chain assignments.
}
\examples{
contig <- clonotype(contig, threshold = 2, levenshtein = F)
           contig <- test.result(contig, threshold = 2, levenshtein = F)
}
\keyword{checks}
