% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjogren_functions.R
\name{volcano}
\alias{volcano}
\title{A function to make volcano plots based on multiple marker dfs}
\usage{
volcano(marker_list, display_n)
}
\arguments{
\item{marker_list}{A named list of the data.frame outputs of the Seurat FindMarkers(), FindAllMarkers() or FindConservedMarkers() function.}

\item{display_n}{An integer stating the amount of (significant) gene names that should be displayed on each side of the volcano plot.}
}
\value{
A list of ggplot objects with the volcano plots, as well as a figure displaying them side-by-side.
}
\description{
This function produces a patchwork of volcano plots based on a list of markers found by running the Seurat FindMarkers() function multiple times.
For example, this function can be used to plot multiple volcano plots side-by-side as a result of comparing two groups of cells within a cluster, for all clusters.
The thresholds for the adjusted p value is set to 0.05, the threshold for the avg_log2FC is set to 0.25.
}
\examples{
markers_1 <- FindMarkers(ident.1, ident.2)
markers_2 <- FindMarkers(ident.3, ident.4)
marker_list <- list(markers_1, markers_2)
plot <- volcano(marker_list, display_n = 10)

}
\keyword{ggplot}
\keyword{marker,}
\keyword{seurat,}
\keyword{volcano,}
